Add array to drop down and implement code to convert currency and grade calc

+++ app/src/main/java/com/example/calculatorproject/MainActivity.java
@@ -27,9 +27,9 @@ public class MainActivity extends AppCompatActivity {
         EditText number2ET = findViewById(R.id.num2ET);
         TextView numberSumTV = findViewById(R.id.resultTV);
 
-        int num1 = Integer.parseInt((number1ET.getText().toString()));
-        int num2 = Integer.parseInt((number2ET.getText().toString()));
-        int sum = num1 + num2;
+        double num1 = Integer.parseInt((number1ET.getText().toString()));
+        double num2 = Integer.parseInt((number2ET.getText().toString()));
+        double sum = num1 + num2;
 
         numberSumTV.setText("" + sum);
     }
@@ -39,9 +39,9 @@ public class MainActivity extends AppCompatActivity {
         EditText number2ET = findViewById(R.id.num2ET);
         TextView numberSumTV = findViewById(R.id.resultTV);
 
-        int num1 = Integer.parseInt((number1ET.getText().toString()));
-        int num2 = Integer.parseInt((number2ET.getText().toString()));
-        int sum = num1 - num2;
+        double num1 = Integer.parseInt((number1ET.getText().toString()));
+        double num2 = Integer.parseInt((number2ET.getText().toString()));
+        double sum = num1 - num2;
 
         numberSumTV.setText("" + sum);
     }
@@ -51,9 +51,9 @@ public class MainActivity extends AppCompatActivity {
         EditText number2ET = findViewById(R.id.num2ET);
         TextView numberSumTV = findViewById(R.id.resultTV);
 
-        int num1 = Integer.parseInt((number1ET.getText().toString()));
-        int num2 = Integer.parseInt((number2ET.getText().toString()));
-        int sum = num1 / num2;
+        double num1 = Integer.parseInt((number1ET.getText().toString()));
+        double num2 = Integer.parseInt((number2ET.getText().toString()));
+        double sum = num1 / num2;
 
         numberSumTV.setText("" + sum);
     }
@@ -63,9 +63,9 @@ public class MainActivity extends AppCompatActivity {
         EditText number2ET = findViewById(R.id.num2ET);
         TextView numberSumTV = findViewById(R.id.resultTV);
 
-        int num1 = Integer.parseInt((number1ET.getText().toString()));
-        int num2 = Integer.parseInt((number2ET.getText().toString()));
-        int sum = num1 * num2;
+        double num1 = Integer.parseInt((number1ET.getText().toString()));
+        double num2 = Integer.parseInt((number2ET.getText().toString()));
+        double sum = num1 * num2;
 
         numberSumTV.setText("" + sum);
     }
@@ -76,9 +76,9 @@ public class MainActivity extends AppCompatActivity {
         EditText number2ET = findViewById(R.id.num2ET);
         TextView numberSumTV = findViewById(R.id.resultTV);
 
-        int num1 = Integer.parseInt((number1ET.getText().toString()));
-        int num2 = Integer.parseInt((number2ET.getText().toString()));
-        int sum = (int) Math.pow(num1, num2);
+        double num1 = Integer.parseInt((number1ET.getText().toString()));
+        double num2 = Integer.parseInt((number2ET.getText().toString()));
+        double sum = Math.pow(num1, num2);
 
         numberSumTV.setText("" + sum);
     }
